Version 2.1.2 (release)

+++ make_web.sh
@@ -1,11 +1,22 @@
 #!/bin/sh -x
 #
-# Move the build web program to the website.  This assumes
-# "flutter build web" has successfully been run on both jrpn15
-# and jrpn16.
+# Move the build web program to the website.  
 #
 
 cd `dirname $0`
+cd jrpn15
+flutter clean
+flutter build web --release
+if [ $? != 0 ]; then
+    exit 1
+fi
+cd ../jrpn16
+flutter clean
+flutter build web --release
+if [ $? != 0 ]; then
+    exit 1
+fi
+cd ..
 rm -rf docs/run docs/run15
 mv jrpn15/build/web docs/run15
 mv jrpn16/build/web docs/run
